+++ setup.cfg
@@ -12,11 +12,11 @@ python_requires = >=3.8
 
 [options.entry_points]
 console_scripts =
-    sort-cspell-json = pre_commit_hooks.format_cspell_json:main
+    sort-cspell-json = format_cspell_json.format_cspell_json:main
 
 [flake8]
 # Modified from https://github.com/ament/ament_lint/blob/ebd524bb9973d5ec1dc48a670ce54f958a5a0243/ament_flake8/ament_flake8/configuration/ament_flake8.ini
-application_import_names = pre_commit_hooks
+application_import_names = format_cspell_json
 extend-ignore = B902,C816,D100,D101,D102,D103,D104,D105,D106,D107,D203,D212,D404,I202,CNL100,E203,E501,Q000
 import-order-style = google
 max-line-length = 100
